Refactor: move node-related functions from PageService to NodeHelper
