+++ lib/Fs/NodeHelper.php
@@ -2,6 +2,7 @@
 
 namespace OCA\Collectives\Fs;
 
+use OCA\Collectives\Model\PageInfo;
 use OCA\Collectives\Service\NotFoundException;
 use OCA\Collectives\Service\NotPermittedException;
 use OCP\Files\File;
@@ -144,5 +145,140 @@ class NodeHelper {
 		}
 
 		return $content;
+	}
+
+	/**
+	 * @param File $file
+	 *
+	 * @return bool
+	 */
+	public static function isPage(File $file): bool {
+		$name = $file->getName();
+		$length = strlen(PageInfo::SUFFIX);
+		return (substr($name, -$length) === PageInfo::SUFFIX);
+	}
+
+	/**
+	 * @param File $file
+	 *
+	 * @return bool
+	 */
+	public static function isLandingPage(File $file): bool {
+		$internalPath = $file->getInternalPath();
+		return ($internalPath === PageInfo::INDEX_PAGE_TITLE . PageInfo::SUFFIX);
+	}
+
+	/**
+	 * @param File $file
+	 *
+	 * @return bool
+	 */
+	public static function isIndexPage(File $file): bool {
+		$name = $file->getName();
+		return ($name === PageInfo::INDEX_PAGE_TITLE . PageInfo::SUFFIX);
+	}
+
+	/**
+	 * @param File $file
+	 *
+	 * @return bool
+	 */
+	public static function indexPageHasOtherContent(File $file): bool {
+		try {
+			foreach ($file->getParent()->getDirectoryListing() as $node) {
+				// True if page and not index page
+				if ($node instanceof File
+					&& !self::isIndexPage($node)
+					&& self::isPage($node)) {
+					return true;
+				}
+				// True if not index page or corresponding attachments folder
+				if (!($node instanceof File && self::isIndexPage($node))
+					&& $node->getName() !== '.attachments.' . $file->getId()) {
+					return true;
+				}
+			}
+		} catch (FilesNotFoundException $e) {
+		}
+
+		return false;
+	}
+
+	/**
+	 * @param Folder $folder
+	 *
+	 * @return bool
+	 */
+	public static function folderHasSubPages(Folder $folder): bool {
+		try {
+			foreach ($folder->getDirectoryListing() as $node) {
+				if ($node instanceof File &&
+					self::isPage($node) &&
+					!self::isIndexPage($node)) {
+					return true;
+				}
+
+				if ($node instanceof Folder) {
+					return self::folderHasSubPages($node);
+				}
+			}
+		} catch (FilesNotFoundException $e) {
+		}
+
+		return false;
+	}
+
+	/**
+	 * @param Folder $folder
+	 * @param string $title
+	 *
+	 * @return int
+	 */
+	public static function folderHasSubPage(Folder $folder, string $title): int {
+		try {
+			foreach ($folder->getDirectoryListing() as $node) {
+				if ($node instanceof File &&
+					strcmp($node->getName(), $title . PageInfo::SUFFIX) === 0) {
+					return 1;
+				}
+
+				if ($node instanceof Folder &&
+					strcmp($node->getName(), $title) === 0 &&
+					$node->nodeExists(PageInfo::INDEX_PAGE_TITLE . PageInfo::SUFFIX)) {
+					return 2;
+				}
+			}
+		} catch (FilesNotFoundException $e) {
+		}
+
+		return 0;
+	}
+
+	/**
+	 * @param Folder $folder
+	 *
+	 * @throws NotFoundException
+	 * @throws NotPermittedException
+	 */
+	public static function revertSubFolders(Folder $folder): void {
+		try {
+			foreach ($folder->getDirectoryListing() as $node) {
+				if ($node instanceof Folder) {
+					self::revertSubFolders($node);
+				} elseif ($node instanceof File) {
+					// Move index page without subpages into the parent folder (if's not the landing page)
+					if (self::isIndexPage($node) && !self::isLandingPage($node) && !self::indexPageHasOtherContent($node)) {
+						$filename = self::generateFilename($folder, $folder->getName(), PageInfo::SUFFIX);
+						$node->move($folder->getParent()->getPath() . '/' . $filename . PageInfo::SUFFIX);
+						$folder->delete();
+						break;
+					}
+				}
+			}
+		} catch (FilesNotFoundException | InvalidPathException $e) {
+			throw new NotFoundException($e->getMessage(), 0, $e);
+		} catch (FilesNotPermittedException | LockedException $e) {
+			throw new NotPermittedException($e->getMessage(), 0, $e);
+		}
 	}
 }
